Made RobotAvatar code cleaner and more readable

+++ RobotCSimulator/Source/ROBOTC/ROBOTCtoC++/DefineGlobalOnce.h
@@ -10,7 +10,7 @@
 // Pure RobotC
 std::array<short, 16> vexRT;
 std::array<short, 10> motor;
-std::array<short, 20> SensorValue;
+std::array<float, 20> SensorValue;
 
 // My variables
 std::array<std::unique_ptr<MotorInfo>, 10> motorInfo;
